Improvement for pasting
This changes allow the cursor to be at the same position while pasting
in normal mode

This improves pasting because we do not need to parcours the whole
line to be where we would like to edit.

An exemple in C where it useful is when we want to copy, paste and modify
some headers.